# docs/swagger.yaml
basePath: /api-market-testnet
definitions:
...
  request.BroadcastTxReq:
    properties:
      txHex:
        type: string
    type: object
...
  respond.BroadcastTxResp:
    properties:
      txId:
        type: string
    type: object
...
  respond.Mrc20BalanceInfo:
    properties:
      balance:
        type: string
      decimals:
        type: string
      mrc20Id:
        type: string
      tick:
        type: string
      tokenName:
        type: string
    type: object
  respond.Mrc20BalanceInfoResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.Mrc20BalanceInfo'
        type: array
      total:
        type: integer
    type: object
...
  respond.Mrc20Info:
    properties:
      amount:
        type: string
      decimals:
        type: string
      mrc20Id:
        type: string
      tick:
        type: string
      txPoint:
        type: string
    type: object
...
  respond.Mrc20ShovelResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.ShovelInfo'
        type: array
      total:
        type: integer
    type: object
  respond.Mrc20TickInfo:
    properties:
      amtPerMint:
        type: string
      beginHeight:
        type: string
      change24h:
        type: string
      decimals:
        type: string
      deployTime:
        type: integer
      deployerAddress:
        type: string
      deployerMetaId:
        type: string
      deployerUserInfo:
        $ref: '#/definitions/respond.UserInfo'
      endHeight:
        type: string
      floorPrice:
        type: string
      floorPriceUsd:
        type: string
      holders:
        type: integer
      marketCap:
        type: string
      marketCapUsd:
        type: string
      metaData:
        type: string
      mintCount:
        type: string
      mintable:
        type: boolean
      mrc20Id:
        type: string
      payCheck: {}
      pinCheck: {}
      pinNumber:
        type: integer
      premineCount:
        type: string
      price:
        type: string
      priceUsd:
        type: string
      qual: {}
      remaining:
        type: string
      supply:
        type: string
      tick:
        type: string
      tokenName:
        type: string
      totalMinted:
        type: string
      totalSupply:
        type: string
      totalVolume:
        type: integer
      txCount:
        type: integer
      type:
        type: string
    type: object
  respond.Mrc20TickListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.Mrc20TickInfo'
        type: array
      total:
        type: integer
    type: object
  respond.Mrc20TickMarketPriceResp:
    properties:
      decimals:
        type: integer
      floorPrice:
        type: string
      floorPriceUsd:
        type: string
      lastPrice:
        type: string
      lastPriceUsd:
        type: string
      marketCap:
        type: string
      marketCapUsd:
        type: string
      price:
        type: string
      priceUsd:
        type: string
      supply:
        type: string
      tick:
        type: string
      tickId:
        type: string
      tokenName:
        type: string
      totalVolume:
        type: integer
    type: object
...
  respond.Mrc20Utxo:
    properties:
      address:
        type: string
      blockHeight:
        type: integer
      chain:
        type: string
      mrc20s:
        items:
          $ref: '#/definitions/respond.Mrc20Info'
        type: array
      orderId:
        type: string
      outputIndex:
        type: integer
      satoshi:
        type: integer
      satoshis:
        type: integer
      scriptPk:
        type: string
      tag:
        description: id-coins
        type: string
      timestamp:
        type: integer
      txId:
        type: string
      vout:
        type: integer
    type: object
  respond.Mrc20UtxoResp:
    properties:
      list:
        items:
          $ref: '#/definitions/respond.Mrc20Utxo'
        type: array
      total:
        type: integer
    type: object
...
  respond.ShovelInfo:
    properties:
      address:
        type: string
      chainName:
        type: string
      content:
        type: string
      contentBody:
        type: string
      contentLength:
        type: integer
      contentSummary:
        type: string
      contentType:
        type: string
      contentTypeDetect:
        type: string
      creator:
        type: string
      dataValue:
        type: integer
      encryption:
        type: string
      genesisFee:
        type: integer
      genesisHeight:
        type: integer
      genesisTransaction:
        type: string
      id:
        type: string
      initialOwner:
        type: string
      isTransfered:
        type: boolean
      location:
        type: string
      metaid:
        type: string
      mrc20MintPin:
        type: string
      mrc20Minted:
        type: boolean
      number:
        type: integer
      offset:
        type: integer
      operation:
        type: string
      originalId:
        type: string
      originalPath:
        type: string
      output:
        type: string
      outputValue:
        type: integer
      parentPath:
        type: string
      path:
        type: string
      pop:
        type: string
      popLv:
        type: integer
      preview:
        type: string
      status:
        type: integer
      timestamp:
        type: integer
      txInIndex:
        type: integer
      txIndex:
        type: integer
      version:
        type: string
    type: object
...
  respond.UserInfo:
    properties:
      avatar:
        type: string
      name:
        type: string
    type: object
info:
  contact:
    name: API Support
  description: MetaID-Market API Service
  title: MetaID-Market API Service
  version: "1.0"
paths:
...
  /api/v1/common/mrc20/address/balance-list:
    get:
      description: Fetch address mrc20 balances
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20BalanceInfoResp'
      summary: Fetch address mrc20 balances
      tags:
      - Common
  /api/v1/common/mrc20/address/shovel:
    get:
      description: Fetch address mrc20 shovels
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20ShovelResp'
      summary: Fetch address mrc20 shovels
      tags:
      - Common
  /api/v1/common/mrc20/address/utxo:
    get:
      description: Fetch address mrc20 utxo
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20UtxoResp'
      summary: Fetch address mrc20 utxo
      tags:
      - Common
  /api/v1/common/mrc20/id-coins/address/utxo:
    get:
      description: Fetch address mrc20 id-coins utxo
      parameters:
      - description: address
        in: query
        name: address
        required: true
        type: string
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20UtxoResp'
      summary: Fetch address mrc20 id-coins utxo
      tags:
      - Common
  /api/v1/common/mrc20/market/price/info:
    get:
      description: Fetch mrc20 market price info
      parameters:
      - description: tickId
        in: query
        name: tickId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TickMarketPriceResp'
      summary: Fetch mrc20 market price info
      tags:
      - Common
  /api/v1/common/mrc20/tick/info:
    get:
      description: Fetch mrc20 tick info
      parameters:
      - description: tickId
        in: query
        name: tickId
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TickInfo'
      summary: Fetch mrc20 tick info
      tags:
      - Common
  /api/v1/common/mrc20/tick/info-list:
    get:
      description: Fetch mrc20 tick info list
      parameters:
      - description: pinnumber/totalminted/holders/txcount/change24H/lastPrice/marketCap/totalSupply
        in: query
        name: orderBy
        type: string
      - description: searchTick
        in: query
        name: searchTick
        type: string
      - description: -1/1, if orderBy is 'change24H/lastPrice/marketCap', it works,
          default '-1'
        in: query
        name: sortType
        type: integer
      - description: true/false/null, default null
        in: query
        name: completed
        type: boolean
      - description: cursor
        in: query
        name: cursor
        type: integer
      - description: size：max=50
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Mrc20TickListResp'
      summary: Fetch mrc20 tick info list
      tags:
      - Common
  /api/v1/common/tx/broadcast:
    post:
      description: Broadcast Tx
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.BroadcastTxReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BroadcastTxResp'
      summary: Broadcast Tx
      tags:
      - Common
...
schemes:
- https
swagger: "2.0"